refactor(tui): split surface_cli_tui_app into focused mixin modules
Extract QwenTuiApp (679 lines) into 4 focused surface-layer mixins:

- surface_cli_tui_compose.py   : compose() UI tree + lifecycle hooks
- surface_cli_tui_handlers.py  : event callbacks + keyboard action_*
- surface_cli_tui_workers.py   : @work threads (slot, login, session)
- surface_cli_tui_utils.py     : table, metrics, status badge, _log_msg

surface_cli_tui_app.py reduced to thin composition root (127 lines)
that inherits all four mixins. surface_cli_tui_css.py added to tracking.

All files pass py_compile, ruff, and mypy checks. AES surface-layer
naming and import rules preserved throughout.
